+++ stellargraph/__init__.py
@@ -16,6 +16,9 @@ __all__ = [
 # Version
 from .version import __version__
 
+# Import modules
+from stellargraph import mapper, layer, utils
+
 # Top-level imports
 from stellargraph.core.graph import StellarGraph, StellarDiGraph
 from stellargraph.core.schema import GraphSchema
@@ -25,3 +28,11 @@ from stellargraph.utils.calibration import (
     expected_calibration_error,
 )
 from stellargraph.utils.ensemble import Ensemble
+
+# Custom layers for keras deserialization:
+custom_keras_layers = {
+    "GraphConvolution": layer.GraphConvolution,
+    "GraphAttention": layer.GraphAttention,
+    "GraphAttentionSparse": layer.GraphAttentionSparse,
+    "SqueezedSparseConversion": layer.SqueezedSparseConversion,
+}