+++ stellargraph/core/utils.py
@@ -123,12 +123,13 @@ def chebyshev_polynomial(X, k):
     return T_k
 
 
-def GCN_Aadj_feats_op(features, A, k=1, **kwargs):
+def GCN_Aadj_feats_op(features, A, k=1, method="gcn"):
+
     """
     This function applies the matrix transformations on the adjacency matrix, which are required by GCN.
     GCN requires that the input adjacency matrix should be symmetric, with self-loops, and normalized.
-    The features and adjacency matrix will be manipulated by either 'localpool', 'chebyshev', or
-    'smoothed' filters.
+    The features and adjacency matrix will be manipulated by either 'gcn' (applying localpool filter as a default), 'chebyshev', or
+    'sgcn' filters.
 
     For more information about 'localpool', 'chebyshev', and 'smoothed' filters, please read details:
         [1] https://en.wikipedia.org/wiki/Chebyshev_filter
@@ -138,44 +139,52 @@ def GCN_Aadj_feats_op(features, A, k=1, **kwargs):
     Args:
         features: node features in the graph
         A: adjacency matrix
-        k (int or None): If filter is 'smoothed' then it should be an integer indicating the power to raise the
+        k (int or None): If method is 'sgcn' then it should be an integer indicating the power to raise the
         normalised adjacency matrix with self loops before multiplying the node features matrix.
-        kwargs: additional arguments for choosing filter: localpool, chebyshev, or smoothed
-                (For example, pass filter="localpool" as an additional argument to apply the localpool filter)
+        If method is 'chebyshev' then it should be an integer indicating the maximum order of the Chebyshev polynomials.
+        method: to specify the filter to use with gcn. If method=gcn, default filter is localpool, other options are 'chebyshev' and 'sgcn'.
+
 
     Returns:
         features (transformed in case of "chebyshev" filter applied), transformed adjacency matrix
     """
 
     def preprocess_adj(adj, symmetric=True):
-        adj = adj + sp.eye(adj.shape[0])
+        adj = adj + sp.diags(np.ones(adj.shape[0]) - adj.diagonal())
         adj = normalize_adj(adj, symmetric)
         return adj
 
     # build symmetric adjacency matrix
     A = A + A.T.multiply(A.T > A) - A.multiply(A.T > A)
-    filter = kwargs.get("filter", "localpool")
 
-    if filter == "localpool":
-        """ Local pooling filters (see 'renormalization trick' in Kipf & Welling, arXiv 2016) """
-        print("Using local pooling filters...")
+    if method == "gcn":
+        # Local pooling filters (see 'renormalization trick' in Kipf & Welling, arXiv 2016) """
+        print("Using GCN (local pooling) filters...")
         A = preprocess_adj(A)
-    elif filter == "chebyshev":
-        """ Chebyshev polynomial basis filters (Defferard et al., NIPS 2016)  """
-        print("Using Chebyshev polynomial basis filters...")
-        max_degree = kwargs.get("max_degree", 2)
-        T_k = chebyshev_polynomial(
-            rescale_laplacian(normalized_laplacian(A)), max_degree
-        )
-        features = [features] + T_k
-    elif filter == "smoothed":
+
+    elif method == "chebyshev":
+        # Chebyshev polynomial basis filters (Defferard et al., NIPS 2016)
+        print("Using Chebyshev polynomial filters...")
+        if isinstance(k, int) and k >= 2:
+            # default minimum degree for Chebyshev polynomials.
+            T_k = chebyshev_polynomial(rescale_laplacian(normalized_laplacian(A)), k)
+            features = [features] + T_k
+        else:
+            raise ValueError(
+                "max_degree should be positive integer of value at least 2 for method='chebyshev'; but received type {} with value {}.".format(
+                    type(k), k
+                )
+            )
+
+    elif method == "sgc":
+        # Smoothing filter (Simplifying Graph Convolutional Networks)
         if isinstance(k, int) and k > 0:
             print("Calculating {}-th power of normalized A...".format(k))
             A = preprocess_adj(A)
             A = A ** k  # return scipy.sparse.csr_matrix
         else:
             raise ValueError(
-                "k should be positive integer for filter='smoothed'; but received type {} with value {}.".format(
+                "k should be positive integer for method='sgcn'; but received type {} with value {}.".format(
                     type(k), k
                 )
             )