+++ tests/layer/test_gcn.py
@@ -28,6 +28,7 @@ import networkx as nx
 import pandas as pd
 import numpy as np
 import keras
+import pytest
 
 
 def create_graph_features():
@@ -43,7 +44,6 @@ def test_GraphConvolution_config():
     conf = gcn_layer.get_config()
 
     assert conf["units"] == 16
-    assert conf["support"] == 1
     assert conf["activation"] == "linear"
     assert conf["use_bias"] == True
     assert conf["kernel_initializer"]["class_name"] == "VarianceScaling"
@@ -57,26 +57,79 @@ def test_GraphConvolution_config():
 
 
 def test_GraphConvolution_init():
-    gcn_layer = GraphConvolution(units=16, support=2, activation="relu")
+    gcn_layer = GraphConvolution(units=16, activation="relu")
 
     assert gcn_layer.units == 16
-    assert gcn_layer.support == 2
     assert gcn_layer.use_bias == True
     assert gcn_layer.get_config()["activation"] == "relu"
 
 
-def test_GraphConvolution_apply():
+def test_GraphConvolution_dense():
     G, features = create_graph_features()
 
-    x_in = Input(shape=(features.shape[1],))
-    A = [Input(shape=(None, None), batch_shape=(None, None), sparse=True)]
-    out = GraphConvolution(2, 1)([x_in] + A)
+    # We need to specify the batch shape as one for the GraphConvolutional logic to work
+    x_t = Input(batch_shape=(1,) + features.shape, name="X")
+    A_t = Input(batch_shape=(1, 3, 3), name="A")
+    output_indices_t = Input(batch_shape=(1, None), dtype="int32", name="outind")
 
-    adj = nx.adjacency_matrix(G)
+    # Note we add a batch dimension of 1 to model inputs
+    adj = nx.to_numpy_array(G)[None, :, :]
+    out_indices = np.array([[0, 1]], dtype="int32")
+    x = features[None, :, :]
 
-    model = keras.Model(inputs=[x_in] + A, outputs=out)
-    preds = model.predict([features, adj], batch_size=adj.shape[0])
-    assert preds.shape == (3, 2)
+    # For dense matrix, remove batch dimension
+    A_mat = Lambda(lambda A: K.squeeze(A, 0))(A_t)
+
+    # Test with final_layer=False
+    out = GraphConvolution(2, final_layer=False)([x_t, output_indices_t, A_mat])
+    model = keras.Model(inputs=[x_t, A_t, output_indices_t], outputs=out)
+    preds = model.predict([x, adj, out_indices], batch_size=1)
+    assert preds.shape == (1, 3, 2)
+
+    # Now try with final_layer=True
+    out = GraphConvolution(2, final_layer=True)([x_t, output_indices_t, A_mat])
+    model = keras.Model(inputs=[x_t, A_t, output_indices_t], outputs=out)
+    preds = model.predict([x, adj, out_indices], batch_size=1)
+    assert preds.shape == (1, 2, 2)
+
+    # Check for errors with batch size != 1
+    # We need to specify the batch shape as one for the GraphConvolutional logic to work
+    x_t = Input(batch_shape=(2,) + features.shape)
+    output_indices_t = Input(batch_shape=(2, None), dtype="int32")
+    with pytest.raises(ValueError):
+        out = GraphConvolution(2)([x_t, A_t, output_indices_t])
+
+
+def test_GraphConvolution_sparse():
+    G, features = create_graph_features()
+    n_nodes = features.shape[0]
+
+    # We need to specify the batch shape as one for the GraphConvolutional logic to work
+    x_t = Input(batch_shape=(1,) + features.shape)
+    A_ind = Input(batch_shape=(1, None, 2), dtype="int64")
+    A_val = Input(batch_shape=(1, None), dtype="float32")
+    output_indices_t = Input(batch_shape=(1, None), dtype="int32")
+
+    # Test with final_layer=False
+    A_mat = SqueezedSparseConversion(shape=(n_nodes, n_nodes))([A_ind, A_val])
+    out = GraphConvolution(2, final_layer=False)([x_t, output_indices_t, A_mat])
+
+    # Note we add a batch dimension of 1 to model inputs
+    adj = nx.to_scipy_sparse_matrix(G, format="coo")
+    A_indices = np.expand_dims(np.hstack((adj.row[:, None], adj.col[:, None])), 0)
+    A_values = np.expand_dims(adj.data, 0)
+    out_indices = np.array([[0, 1]], dtype="int32")
+    x = features[None, :, :]
+
+    model = keras.Model(inputs=[x_t, output_indices_t, A_ind, A_val], outputs=out)
+    preds = model.predict([x, out_indices, A_indices, A_values], batch_size=1)
+    assert preds.shape == (1, 3, 2)
+
+    # Now try with final_layer=True
+    out = GraphConvolution(2, final_layer=True)([x_t, output_indices_t, A_mat])
+    model = keras.Model(inputs=[x_t, output_indices_t, A_ind, A_val], outputs=out)
+    preds = model.predict([x, out_indices, A_indices, A_values], batch_size=1)
+    assert preds.shape == (1, 2, 2)
 
 
 def test_GCN_init():
@@ -95,21 +148,59 @@ def test_GCN_init():
     assert gcnModel.dropout == 0.5
 
 
-def test_GCN_apply():
+def test_GCN_apply_dense():
     G, features = create_graph_features()
-    adj = nx.adjacency_matrix(G)
+    adj = nx.to_numpy_array(G)[None, :, :]
+    n_nodes = features.shape[0]
 
     nodes = G.nodes()
     node_features = pd.DataFrame.from_dict(
         {n: f for n, f in zip(nodes, features)}, orient="index"
     )
-    G = StellarGraph(G, node_type_name="node", node_features=node_features)
+    G = StellarGraph(G, node_features=node_features)
 
-    generator = FullBatchNodeGenerator(G)
+    generator = FullBatchNodeGenerator(G, sparse=False, method="none")
     gcnModel = GCN([2], ["relu"], generator=generator, dropout=0.5)
 
     x_in, x_out = gcnModel.node_model()
     model = keras.Model(inputs=x_in, outputs=x_out)
-    preds = model.predict([features, adj], batch_size=adj.shape[0])
 
-    assert preds.shape == (3, 2)
+    # Check fit method
+    out_indices = np.array([[0, 1]], dtype="int32")
+    preds_1 = model.predict([features[None, :, :], out_indices, adj])
+    assert preds_1.shape == (1, 2, 2)
+
+    # Check fit_generator method
+    preds_2 = model.predict_generator(generator.flow(["a", "b"]))
+    assert preds_2.shape == (1, 2, 2)
+
+    assert preds_1 == pytest.approx(preds_2)
+
+
+def test_GCN_apply_sparse():
+    G, features = create_graph_features()
+    adj = nx.to_numpy_array(G)[None, :, :]
+    n_nodes = features.shape[0]
+
+    nodes = G.nodes()
+    node_features = pd.DataFrame.from_dict(
+        {n: f for n, f in zip(nodes, features)}, orient="index"
+    )
+    G = StellarGraph(G, node_features=node_features)
+
+    generator = FullBatchNodeGenerator(G, sparse=False, method="none")
+    gcnModel = GCN([2], ["relu"], generator=generator, dropout=0.5)
+
+    x_in, x_out = gcnModel.node_model()
+    model = keras.Model(inputs=x_in, outputs=x_out)
+
+    # Check fit method
+    out_indices = np.array([[0, 1]], dtype="int32")
+    preds_1 = model.predict([features[None, :, :], out_indices, adj])
+    assert preds_1.shape == (1, 2, 2)
+
+    # Check fit_generator method
+    preds_2 = model.predict_generator(generator.flow(["a", "b"]))
+    assert preds_2.shape == (1, 2, 2)
+
+    assert preds_1 == pytest.approx(preds_2)